Move answer and classification to top level attributes, rename higherlevel (#12)
* Move answer and classification to top level attributes

* New namespaces for answers and classifications

* Meant to make the method create

* Go up the class stack since we don't need all the things that engineapiresource gives us
