+++ openai/api_resources/__init__.py
@@ -4,7 +4,8 @@ from openai.api_resources.engine import Engine
 from openai.api_resources.error_object import ErrorObject
 from openai.api_resources.event import Event
 from openai.api_resources.file import File
-from openai.api_resources.higherlevel import HigherLevel
+from openai.api_resources.answer import Answer
+from openai.api_resources.classification import Classification
 from openai.api_resources.plan import Plan
 from openai.api_resources.run import Run
 from openai.api_resources.snapshot import Snapshot
